feat: подключить отправку форм

+++ src/scripts/utils/Api.js
@@ -55,19 +55,18 @@ class Api {
   }
 
   //поставить лайк
-  putLike(id){
-    return fetch(`${this._url}/cards/${id}/likes`, {
-      method: 'PUT',
-      headers: this._headers
-    }).then(this._errorHandler)
-  }
-
-  //удалить лайк
-  deleteLike(id){
-    return fetch(`${this._url}/cards/${id}/likes`, {
-      method: 'DELETE',
-      headers: this._headers
-    }).then(this._errorHandler)
+  changeLikeCardStatus(id, isNotLiked){
+    if (isNotLiked) {
+      return fetch(`${this._url}/cards/${id}/likes`, {
+        method: 'PUT',
+        headers: this._headers
+      }).then(this._errorHandler)
+    } else {
+      return fetch(`${this._url}/cards/${id}/likes`, {
+        method: 'DELETE',
+        headers: this._headers
+      }).then(this._errorHandler)
+    }
   }
 
   //обновить аватар пользователя (PATCH)
